refactor: 게시글 수정 기능 리팩토링
