+++ apps/web/src/6_shared/ui/ImageUploader.tsx
@@ -1,49 +1,23 @@
 'use client';
-
-import { compressImage, COMPRESSION_OPTIONS } from '@/6_shared/utils/optimizeImage';
 import { cn } from '@/6_shared/utils/style';
-import { forwardRef } from 'react';
+import Image from 'next/image';
+import galleryIcon from 'public/icon/icon-gallery.svg';
 
-interface ImageUploaderProps {
-    setImage: (image: File | null) => void;
-    preview?: React.ReactNode;
-    setPreviewUrl: (url: string | null) => void;
+interface ImageUploaderProps extends React.InputHTMLAttributes<HTMLInputElement> {
     className?: string;
 }
 
-const ImageUploader = forwardRef<HTMLInputElement, ImageUploaderProps>(
-    ({ setImage, preview, setPreviewUrl, className }, ref) => {
-        const handleUploadFile = async (e: React.ChangeEvent<HTMLInputElement>) => {
-            const file = e.target.files?.[0];
-            if (file) {
-                const optimizedImage = await compressImage(file, COMPRESSION_OPTIONS.THUMBNAIL);
-                setImage(optimizedImage);
-                setPreviewUrl(URL.createObjectURL(optimizedImage));
-            }
-        };
-
-        return (
-            <>
-                <div className={cn('relative', className)}>
-                    <input
-                        id='image-uploader'
-                        type='file'
-                        accept='image/*'
-                        className='hidden'
-                        onChange={handleUploadFile}
-                        multiple={false}
-                        ref={ref}
-                    />
-                    <label className={cn('cursor-pointer size-full')} htmlFor='image-uploader'>
-                        <img src='/icon/icon-gallery.svg' alt='gallery' className='w-full h-full object-cover' />
-                    </label>
-                    {preview}
-                </div>
-            </>
-        );
-    },
-);
-
-ImageUploader.displayName = 'ImageUploader';
+const ImageUploader = ({ className, ...props }: ImageUploaderProps) => {
+    return (
+        <>
+            <div className={cn('relative', className)}>
+                <input id='image-uploader' type='file' accept='image/*' className='hidden' {...props} />
+                <label className={cn('cursor-pointer size-full')} htmlFor='image-uploader'>
+                    <Image src={galleryIcon} alt='gallery' className='w-full h-full object-cover' />
+                </label>
+            </div>
+        </>
+    );
+};
 
 export default ImageUploader;
